feat: services setting

+++ app/services/user_service.py
@@ -1,8 +1,8 @@
+from fastapi import HTTPException
 from sqlalchemy.orm import Session
 from app.models.user import User
-from app.schemas.user import UserCreate, UserUpdate, UserResponse
-
-# nickname, address, src -> Optional
+from app.schemas.user import UserCreate, UserUpdate
+from fastapi.responses import JSONResponse
 
 
 def create_user(db: Session, user: UserCreate):
@@ -20,6 +20,8 @@ def create_user(db: Session, user: UserCreate):
 
 def update_user(db: Session, user_id: int, user: UserUpdate):
     db_user = db.query(User).filter(User.id == user_id).first()
+    if not db_user:
+        raise HTTPException(status_code=404, detail="User not found")
     db_user.nickname = user.nickname
     db_user.address = user.address
     db_user.src = user.src
@@ -30,15 +32,23 @@ def update_user(db: Session, user_id: int, user: UserUpdate):
 
 def delete_user(db: Session, user_id: int):
     db_user = db.query(User).filter(User.id == user_id).first()
+    if not db_user:
+        raise HTTPException(status_code=404, detail="User not found")
     db.delete(db_user)
     db.commit()
-    return {"message": "User deleted successfully"}
+    return JSONResponse(content={"message": "User deleted successfully"}, status_code=200)
 
 
 def get_user_by_id(db: Session, user_id: int):
-    return db.query(User).filter(User.id == user_id).first()
+    user = db.query(User).filter(User.id == user_id).first()
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    return user
 
 
 def get_user_by_email(db: Session, email: str):
-    return db.query(User).filter(User.email == email).first()
+    user = db.query(User).filter(User.email == email).first()
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    return user
 # Compare this snippet from app/api/v1/endpoints/user.py:
